==> Review-Classification-using-NLP/2_hanspell.ipynb <==
## hanspell 이용 맞춤법 띄어쓰기 수정

!pip install git+https://github.com//ssut/py-hanspell.git
    
from hanspell import spell_checker
import pandas as pd

df = pd.read_csv('data/review_result.csv')

review_li = df['review']


i=1
re_review = []
for review in review_li:
    print(review)
    print('-->')
    try:
        spelled_sent = spell_checker.check(review)
        hanspell_sent = spelled_sent.checked
    except:
        print(' ')
    else:
        hanspell_sent=review
    print(hanspell_sent)
    print(i)
    i+=1
    print('=====================================================================')
    re_review.append(hanspell_sent)
    
    
df['re_review']=re_review
df.to_excel('py_hanspell.xlsx',index=False, encoding="utf-8")
