==> Review-Classification-using-NLP/3_labeled.ipynb <==
import pandas as pd

## 맞춤법/띄어쓰기된 데이터 불러오기

df = pd.read_excel('data/py_hanspell.xlsx')

df

## 긍부정 label

df_pos = df[df['rate']==5]
df_neg = df[df['rate']<2.5]

df_pos['is_pos']=1
df_neg['is_pos']=0

df_labeled = pd.concat([df_pos, df_neg])
df_labeled.reset_index(drop=True, inplace=True)

df_labeled.head()

labeled_result = pd.DataFrame()
labeled_result['review'] = df_labeled['hanspell_review']
labeled_result['is_pos'] = df_labeled['is_pos']

labeled_result

## 부정 라벨 리뷰 추가 크롤링

extra_df = pd.read_csv('data/extra_neg_rev.csv', encoding='utf-8')

extra_neg = pd.DataFrame()
extra_neg['review'] = extra_df['review']
extra_neg['is_pos'] = 0

labeled_result = pd.concat([labeled_result, extra_neg])

labeled_result = labeled_result.dropna()

labeled_result

## 데이터 저장하기

labeled_result.to_csv('data/PNlabeled_review.csv',index=False, encoding="utf-8")
