==> Review-Classification-using-NLP/6_train_test_split.ipynb <==
from sklearn.model_selection import train_test_split
import pandas as pd

## 데이터 불러오기

df = pd.read_csv('data/PNlabeled_review.csv', encoding='utf-8')

## train, test 분리하기

train_set,test_set  = train_test_split(df, test_size=0.25, random_state=42)

test_set[test_set['is_pos']==1]

## 데이터 저장하기

train_set.to_csv('data/PNtrain_set.csv', index=None, encoding='utf-8')

test_set.to_csv('data/PNtest_set.csv', index=None, encoding='utf-8')
