==> Review-Classification-using-NLP/7_preprocessing(colab).ipynb <==
# 형태소분석기 관련 설치
!apt-get update
!apt-get install g++ openjdk-8-jdk

!pip install JPype1
!pip install rhinoMorph

import numpy as np
import pandas as pd
import re
import json
import rhinoMorph
from tensorflow.python.keras.preprocessing.sequence import pad_sequences
from tensorflow.python.keras.preprocessing.text import Tokenizer

## 데이터 불러오기

train_set = pd.read_csv('/content/drive/MyDrive/KTO_Project/data/PNtrain_set.csv', encoding='utf-8')
test_set = pd.read_csv('/content/drive/MyDrive/KTO_Project/data/PNtest_set.csv', encoding='utf-8')

train_set.head()

## 데이터 하나 특수문자 처리

review_text = re.sub("[^가-힣ㄱ-ㅎㅏ-ㅣ\\s]", "", train_set['review'][0])
print(review_text)

## 형태소 분석

rn = rhinoMorph.startRhino()

morphed_review = rhinoMorph.onlyMorph_list(rn, review_text, pos=['NNG', 'NNP', 'NP', 'VV', 'VA', 'XR', 'IC', 'MM', 'MAG', 'MAJ'], eomi=True)
print(morphed_review)

## 불용어 사전 이용하여 불용어 제거

stop_word = pd.read_csv('/content/drive/MyDrive/KTO_Project/data/stopword_dict.csv', encoding='utf-8')

stop_word = set(stop_word['stopWords'])

stop_word

clean_review = []
for token in morphed_review:
  if not token in stop_word:
    clean_review.append(token)
print(clean_review)

## 데이터 전처리 함수 작성

def preprocessing(review, rn, stop_words=[]):
  review_text = re.sub("[^가-힣ㄱ-ㅎㅏ-ㅣ\\s]", "", review)

  morphed_review = rhinoMorph.onlyMorph_list(rn, review_text, pos=['NNG', 'NNP', 'NP', 'VV', 'VA', 'XR', 'IC', 'MM', 'MAG', 'MAJ'], eomi=True)
  
  clean_review=[]
  for token in morphed_review:
    if not token in stop_word:
      clean_review.append(token)
  return clean_review

clean_train_review=[]
clean_test_review=[]
for review in train_set['review']:
  clean_train_review.append(preprocessing(review, rn, stop_words=stop_word))
for review in test_set['review']:
  clean_test_review.append(preprocessing(review, rn, stop_words=stop_word))

clean_train_review[:2]

## 인덱스 벡터 변환 및 패딩

tokenizer = Tokenizer()
tokenizer.fit_on_texts(clean_train_review)
train_sequences = tokenizer.texts_to_sequences(clean_train_review)
test_sequences = tokenizer.texts_to_sequences(clean_test_review)

word_vocab = tokenizer.word_index

MAX_SEQUENCE_LENGTH = 14 # EDA 과정에서 단어 평균 개수 14개

train_inputs = pad_sequences(train_sequences, maxlen=MAX_SEQUENCE_LENGTH, padding='post') # 데이터 벡터화
train_labels = np.array(train_set['is_pos']) # 학습 데이터 라벨
test_inputs = pad_sequences(test_sequences, maxlen=MAX_SEQUENCE_LENGTH, padding='post') # 평가 데이터 벡터화
test_labels = np.array(test_set['is_pos']) # 평가 데이터 라벨

## 전처리된 데이터 저장

DATA_IN_PATH = '/content/drive/MyDrive/KTO_Project/data/'
TRAIN_INPUT_DATA = 'nsmc_train_input.npy'
TRAIN_LABEL_DATA = 'nsmc_train_label.npy'
TEST_INPUT_DATA = 'nsmc_test_input.npy'
TEST_LABEL_DATA = 'nsmc_test_label.npy'
DATA_CONFIGS = 'data_configs.json'

data_configs = {}

data_configs['vocab'] = word_vocab
data_configs['vocab_size'] = len(word_vocab)+1

# 전처리된 학습 데이터 넘파이 형태로 저장
np.save(open(DATA_IN_PATH + TRAIN_INPUT_DATA, 'wb'), train_inputs)
np.save(open(DATA_IN_PATH + TRAIN_LABEL_DATA, 'wb'), train_labels)
# 전처리된 평가 데이터 넘파이 형태로 저장
np.save(open(DATA_IN_PATH + TEST_INPUT_DATA, 'wb'), test_inputs)
np.save(open(DATA_IN_PATH + TEST_LABEL_DATA, 'wb'), test_labels)

# 데이터 사전을 json 형태로 저장
json.dump(data_configs, open(DATA_IN_PATH + DATA_CONFIGS, 'w'), ensure_ascii=False)
