==> Review-Classification-using-NLP/10_wordEmbedding.ipynb <==
!pip install gensim

import matplotlib as mpl
import matplotlib.pyplot as plt
import pandas as pd
import os.path
embedding_dim = 200
filepath = '/content/drive/MyDrive/KTO_Project/data/'

# 그래프에서 한글표현을 위해 폰트를 설치
%config InlineBackend.figure_format = 'retina'       # 폰트가 깨끗하게 보이도록 설정
!apt install fonts-nanum                             # 나눔 폰트 설치
#!apt-get -qq -y install fonts-nanum > /dev/null     # 폰트 설치 대안 코드

import matplotlib.font_manager as fm
fontpath = '/usr/share/fonts/truetype/nanum/NanumMyeongjo.ttf'
font = fm.FontProperties(fname=fontpath, size=9)

# 기본 글꼴 변경
mpl.pyplot.rc('font', family='NanumMyeongjo')
mpl.font_manager._rebuild()

# 형태소분석기 관련 설치
!apt-get update
!apt-get install g++ openjdk-8-jdk

!pip install JPype1
!pip install rhinoMorph

## 리뷰 불러오기

df = pd.read_excel('/content/drive/MyDrive/KTO_Project/data/hanspell_review.xlsx')

df

df[df['hanspell_review'].isna()]
df = df.replace('\n', ' ')

#4. RHINO 시작
import rhinoMorph
rn = rhinoMorph.startRhino()

md_df = df['hanspell_review'] 
print(md_df)

## 텍스트 형태소 분석

#Word2Vec – 텍스트의 형태소 분석
total_lines = len(md_df)
md_rev = []
cnt=0
with open(filepath+'morphed.txt', 'w', encoding='utf-8')as f:
  for data_each in md_df:
    morphed_data_each = rhinoMorph.onlyMorph_list(rn, data_each, pos=['NNG', 'NNP', 'NP', 'VV', 'VA', 'XR', 'IC', 'MM', 'MAG', 'MAJ'])
    joined_data_each = ' '.join(morphed_data_each)
    if joined_data_each:
      f.write(joined_data_each + '\n')
    cnt += 1
    if (cnt % 1000) == 0:  # 진행 정도 확인을 위해 1000번째 문장마다 확인
      print(round(cnt/total_lines * 100, 3), '%')
  print('Morphological Analysis Completed.')

## 형태소 분석 결과 읽어 리스트로 만들기 - Word2Vec

def read_data(filename, encoding='utf-8'):
  with open(filename, 'r', encoding=encoding) as f:
    data = [line.split(' ') for line in f.read().splitlines()]
  return data

data = read_data(filepath+'morphed.txt', 'utf-8')

print(len(data))
print(type(data))
print(data[:3])

## 임베딩 구성 - Word2Vec

from gensim.models import Word2Vec
os.chdir(filepath)

model = Word2Vec(sentences=data, size=embedding_dim, window=15, min_count=5, workers=4, sg=1)

model.save('embedding_skipgram.model')

## 임베딩 값 저장 - Word2Vec

words = list(model.wv.vocab)
with open('embedding_skipgram.txt', 'w') as f:
  for word in words:
    data = model.wv[word].tolist()
    print('data_pre: ', data)

    data.insert(0, word)
    print('data_after: ', data)

    for item in data:
      f.write('%s ' %item)
    f.write('\n')

## 유사어 찾기

model = Word2Vec.load('embedding_skipgram.model')

similarWords = model.wv.most_similar(positive=['위생'], topn=100)
print(similarWords)

word=[]
for similarWord in similarWords:
  word.append(similarWord[0])
print(word)

## 두 단어 사이의 유사도 계산 - Word2Vec

print('위생과 청결 유사도: ', model.wv.similarity('입장료', '나빠'))

print('서비스 편의: ', model.wv.similarity('서비스', '편의'))

print('가격 착하: ', model.wv.similarity('가격', '착하'))

model.wv.similarity('대구','부산')
