==> Review-Classification-using-NLP/11_multiLabeled_score.ipynb <==
import pandas as pd

df = pd.read_csv('data/ota_morph.csv', encoding='utf-8')

df = df.dropna()

morphed_list=[]
for item in df['md_review_eomi_t']:
    morphed_list.append(item.split())
df['morphed_list'] = morphed_list

df

clean_dict = ['위생', '청결', '더럽다', '쓰레기', '지저분', '냄새나다', '깔끔', '곰팡이', '엉망', '간지럽다', '청소', '깨끗', 
              '미비', '청경', '이불상','찜찜','찝찝', '물청소', '비위', '물때', '너저분', '위생적', '누렇다','머리카락','부유물', 
              '드럽다', '행주','깨끗이','부스러기','얼룩덜룩','먼지','손톱','걸레질','방상태','상태','침구','위생적','쥐',
             '진드기','구더기','안치우','오물','벌레','끈적끈적','바퀴','구역질','역겹다','음모','털이','가닥','얼룩이','얼룩',
              '털','핏자국','자국','시트','오염','시커멓다','체모','쉬내','이물질','찌꺼기','덕지덕지','끈적거리다','변색',
             '시체','담배꽁초','축축','끈적이다','미끌거리다','미끌다','얼룩지다','휴지통','끈적','닦이다','바퀴벌레','하수구','더러움',
              '거미줄','근질근질','악취','비린내','물비린내','깨끗히']

service_dict = ['마인드','써비스','발레파킹','직원','사장','여사장','인성','종업원','접대','차별','성심성의껏','전문성','다정',
                '절하','자상','친절','친절하다','인자','나긋나긋','친절히','살갑다','상냥','친화력','프로페셔널','유쾌','부지런','주인집',
                '짱짱맨','불친절','무뚝뚝','싸가지','말투','여직원','태도','근무자','사무적','투철','팀장','관리인','요원',
               '커뮤니케이션','프론트','데스크','카운터','능력','접객','클레임','친절히','알바','눈빛','뚱하다','무표정','퉁명','성의',
               '서비스업','툭툭','유니폼','표정','반말','호텔리어','지배인','교육','융통','융통성','쌩','정중히','형식적','대꾸',
                '대답','무시','언성','짜증내다','화내다','찌푸리다','배려','업그레이드','도움','가이드','전문적','공손','질문','해설',
                '세심','사모','매너','체크인','체크아웃','안내']


facility_dict = ['시설물','게임방','인테리','당구장','탁구장','신식','부대시설','신설','인프라','탁구대','최신식',
                 '휘트니스','편리','편안','따뜻','시원','노후','조용','넓다','편히','편하다','불편하다','불편','푹신','낙후',
                 '노후화','낡다','겉보기','보수','부대','시설','편안히','갖추다','어매','어매니티','협소','에어컨',
                 '리모콘','충전기','컴퓨터','편의','샴푸','암막','인테리어','창문','점검','파손','고풍','놀이터','헬스장','조경',
                 '발코니','라운지','로션','단수','리모컨','생활용품','수리','개방감','방음','구비','비치','트리트먼트','찬바람','머신',
                '세안','세면도구','용품','바디워시','샴','샴프','청결제','쉐이빙','파우치','바디로션','종이컵','레인지','전자','전자레인지',
                '드라이어','헤어드라이어','드라이기','전자렌지','렌지','배치','완비','면봉','클렌징','휴게','소음','옆방','건조기','정비',
                 '윗층','아랫층','위층','가스렌지','제빙','고데기','식수','치약','칫솔','스낵바','주차장','전등','실내화','바베큐',
                 '반려동물','반려견', '지하주차장','주차타워','협소','기계식','만차','타워','수압','샤워기','물줄기','온도','배수',
                 '스위치','조명','보일러','온돌방','커튼','공용','외관','데스크탑','분위기','다리미','전기요','가운', '엘리베이터','엘베',
                 '테이블','리모델링','와이파이','고장','텔레비전','콘센트','공기청정기','무인','인터넷','프린터','냉온수','물품','디자인',
                 '좁아터지다','좁다']

safety_dict = ['안전','도어','비밀번호','화재경보기','화재','보안','금고','잠금','경보','대피','락','도어락','예방','외부인','침입',
               '안심','비상약','비상','차단기','치안','도난','바이러스','방역','자물쇠','열쇠','소독','비상구','응급','사이렌','무단','경보',
               '소화전','대피로','소화기','사고']

price_dict = ['비지떡','가성','가성비','가격대비','가격','가격대','혜자','창렬','금액','숙박비','숙박료','개이득','값어치','값',
              '객실료','저렴','합리적','실속','예산','경제적','싸게','비싸다','이득','이용료','좀비','입장료','금액','요금',
             '바가지','특가','프로모션','세일','할인','비대','성비','절감','지출','적정가격']

revisit_dict = ['또','자주','재방문','방문','자주자주','꼭','기약','언젠가','추천','강추','소개','주변인','지인','추전','비추','비추다',
                '비추천','꾸욱','오랜만','오랫만','간만']

label_list=[]
label_score=[]

for review in df['morphed_list']:
    clean_score = 0
    service_score = 0
    facility_score = 0
    safety_score = 0
    price_score = 0
    revisit_score = 0
    
    for item in review:
        for clean in clean_dict:
            if item==clean:
                clean_score += 1
        for service in service_dict:
            if item==service:
                service_score += 1
        for facility in facility_dict:
            if item==facility:
                facility_score += 1
        for safety in safety_dict:
            if item==safety:
                safety_score += 1
        for price in price_dict:
            if item==price:
                price_score += 1
        for revisit in revisit_dict:
            if item==revisit:
                revisit_score += 1
    
# 위생/청결:1, 서비스:2, 시설/편의:3, 안전:4, 가격만족도:5, 재방문/지인추천:6
        
    temp = [0, 0]
    if temp[1]<clean_score:
        temp = [1, clean_score]
    if temp[1]<service_score:
        temp = [2, service_score]    
    if temp[1]<facility_score:
        temp = [3, facility_score]
    if temp[1]<safety_score:
        temp = [4, safety_score]
    if temp[1]<price_score:
        temp = [5, price_score]
    if temp[1]<revisit_score:
        temp = [6, revisit_score]
    
    label_list.append(temp[0])
    label_score.append(temp[1])
df['labeled'] = label_list
df['label_score'] = label_score

cnt = 0
for x in label_list:
    if x == [0,0]:
        cnt += 1
print(cnt)

df

df['label_score'].unique()

df[df['label_score']==2]

df[df['labeled']==6]
